En cours de modif, partage pour resolution du pb de workspace
