+++ Cargo.toml
@@ -12,7 +12,7 @@ embedded-hal = "0.2.4"
 ieee802154   = "0.5.0"
 nb           = "1.0.0"
 fixed        = "1.7.0"
-micromath    = "2.0.0"
+#micromath    = "2.0.0"
 
 [dependencies.serde]
 version = "1.0.125"
@@ -33,3 +33,45 @@ std = ["ssmarshal/std", "serde/std", "num_enum/std"]
 
 [workspace]
 members = ["example_stm32f1", "example_raspberry"]
+
+# cargo build/run
+[profile.dev]
+codegen-units = 1
+debug = 2
+debug-assertions = true # <-
+incremental = false
+opt-level = 3 # <-
+overflow-checks = true # <-
+
+# cargo test
+[profile.test]
+codegen-units = 1
+debug = 2
+debug-assertions = true # <-
+incremental = false
+opt-level = 3 # <-
+overflow-checks = true # <-
+
+# cargo build/run --release
+[profile.release]
+codegen-units = 1
+debug = 2
+debug-assertions = false # <-
+incremental = false
+# NOTE disabled to work around issue rust-lang/rust#90357
+# the bug results in log messages not having location information
+# (the line printed below the log message that contains the file-line location)
+# lto = 'fat'
+opt-level = 3 # <-
+overflow-checks = false # <-
+
+# cargo test --release
+[profile.bench]
+codegen-units = 1
+debug = 2
+debug-assertions = false # <-
+incremental = false
+# see comment in the profile.release section
+lto = 'false'
+opt-level = 3 # <-
+overflow-checks = false # <-
