+++ example_stm32f1/Cargo.toml
@@ -23,51 +23,13 @@ panic-probe = { version = "0.3.0", features = ["print-defmt"] }
 stm32f1xx-hal = { version = "0.8.0", features = ["stm32f103", "rt", "medium"] }
 dw3000 = {path = "../"}
 nb           = "1.0.0"
+embedded-timeout-macros = "0.3.0"
+embedded-hal = "0.2.4"
+
+
 
 [dev-dependencies]
 defmt-test = "0.3.0"
-
-# cargo build/run
-[profile.dev]
-codegen-units = 1
-debug = 2
-debug-assertions = true # <-
-incremental = false
-opt-level = 3 # <-
-overflow-checks = true # <-
-
-# cargo test
-[profile.test]
-codegen-units = 1
-debug = 2
-debug-assertions = true # <-
-incremental = false
-opt-level = 3 # <-
-overflow-checks = true # <-
-
-# cargo build/run --release
-[profile.release]
-codegen-units = 1
-debug = 2
-debug-assertions = false # <-
-incremental = false
-# NOTE disabled to work around issue rust-lang/rust#90357
-# the bug results in log messages not having location information
-# (the line printed below the log message that contains the file-line location)
-# lto = 'fat'
-opt-level = 3 # <-
-overflow-checks = false # <-
-
-# cargo test --release
-[profile.bench]
-codegen-units = 1
-debug = 2
-debug-assertions = false # <-
-incremental = false
-# see comment in the profile.release section
-lto = 'false'
-opt-level = 3 # <-
-overflow-checks = false # <-
 
 # uncomment this to switch from the crates.io version of defmt to its git version
 # check app-template's README for instructions
