+++ example_stm32f1/src/bin/simple_rtt_responder.rs
@@ -2,17 +2,19 @@
 #![no_std]
 
 /*  
-	simple anchor example to be used with simple tag example to perform RTT measurements (simple sided method)
+	simple tag example to be used with simple anchor example to perform RTT measurements (simple sided method)
 
 	SIMPLE SIDED RTT MEASUREMENT TECHNIQUE :
 
-	   TAG					ANCHOR
-	T1	|--------____		  |
-		| 		     -------> |	T2	
-		|					  |						
-		|		  ____--------| T3
-	T4	| <-------
+	INITIATOR				RESPONDER
+	T1	| -------______		   |
+		| 		       ------> |	T2	
+		|					   |						
+		|		 ______------- | T3
+	T4	| <------
 */
+
+use example_stm32f1 as _; // global logger + panicking-behavior + memory layout
 
 use cortex_m_rt::entry;
 use stm32f1xx_hal::{
@@ -21,7 +23,11 @@ use stm32f1xx_hal::{
 	prelude::*,
 	spi::{Mode, Phase, Polarity, Spi},
 };
-use dw3000::{hl, Config};
+use dw3000::{
+	hl,
+	Config,
+	time::Instant,
+};
 use nb::block;
 
 #[entry]
@@ -102,65 +108,53 @@ fn main() -> ! {
 
     defmt::println!("Init OK");
 
-	dw3000.set_antenna_delay(0,0).unwrap();
+	dw3000.set_antenna_delay(4416, 16500).expect("Failed set antenna delay.");
 
 	loop {
 		/**************************** */
-		/******** TRANSMITTER ******* */
-		/**************************** */
-		let mut sending = dw3000
-			.send(
-				&[1, 2, 3, 4, 5],
-				hl::SendTime::Now,
-				Config::default(),
-			)
-			.expect("Failed to configure transmitter");
-		let result = block!(sending.s_wait());
-		let t1:u64 = result.unwrap().value();
-		dw3000 = sending.finish_sending().expect("Failed to finish sending");
-
-		/**************************** */
-		/***** RECEIVER T2 and T3 *** */
+		/********* RECEIVER ********* */
 		/**************************** */
 		let mut receiving = dw3000
 			.receive(Config::default())
 			.expect("Failed configure receiver.");
 		let mut buffer = [0; 1024];
-		let result = block!(receiving.r_wait(&mut buffer));
-
+		let t2 = block!(receiving.r_wait(&mut buffer)).expect("error during the reception").rx_time.value();
 		dw3000 = receiving
 			.finish_receiving()
 			.expect("Failed to finish receiving");
-		let result = result.unwrap();
-		let t4:u64 = result.rx_time.value();
-		let x = result.frame.payload;
-		let t2: u64 = ((x[0] as u64) << (8 * 4)) 
-					+ ((x[1] as u64) << (8 * 3))
-					+ ((x[2] as u64) << (8 * 2))
-					+ ((x[3] as u64) << (8 * 1))
-					+ (x[4] as u64);
-		let t3: u64 = ((x[5] as u64) << (8 * 4)) 
-					+ ((x[6] as u64) << (8 * 3))
-					+ ((x[7] as u64) << (8 * 2))
-					+ ((x[8] as u64) << (8 * 1))
-					+ (x[9] as u64);
 
-		defmt::println!("T1 = {:?} | T2 = {:?}", t1, t2);
-		defmt::println!("T3 = {:?} | T4 = {:?}", t3, t4);
-		defmt::println!("measured distance = {}", calc_distance(t1,t2,t3,t4));
+		// We need to calculate a time (in ticks) at which we want to send the response
+		let response_delay = t2 + (5000 * 63898) as u64; // T2(ticks) + (chosen_delay(µs) * clock_speed)
+		let t3 = ((response_delay >> 9) << 9) + dw3000.get_tx_antenna_delay();  // T3(ticks) = delay(31 MSB) + sending_antenna_delay 
+
+		let response_tab = [
+			((t2 >> (8 * 4) ) & 0xFF ) as u8,
+			((t2 >> (8 * 3) ) & 0xFF ) as u8,
+			((t2 >> (8 * 2) ) & 0xFF ) as u8,
+			((t2 >>  8      ) & 0xFF ) as u8,
+			 (t2              & 0xFF ) as u8,
+			((t3 >> (8 * 4) ) & 0xFF ) as u8,
+			((t3 >> (8 * 3) ) & 0xFF ) as u8,
+			((t3 >> (8 * 2) ) & 0xFF ) as u8,
+			((t3 >>  8      ) & 0xFF ) as u8,
+			 (t3              & 0xFF ) as u8,
+		];
+
+		/**************************** */
+		/*** TRANSMITTER T2 and T3 ** */
+		/**************************** */
+
+		let mut sending = dw3000
+			.send(
+				&response_tab,
+				hl::SendTime::Delayed(Instant::new(response_delay).unwrap()),
+				Config::default(),
+			)
+			.expect("Failed configure transmitter");
+		block!(sending.s_wait()).expect("Error sending");
+		dw3000 = sending.finish_sending().expect("Failed to finish sending");
+
+		defmt::println!("T2 = {:?}", t2);
+		defmt::println!("T3 = {:?}\n", t3);
 	}
-}
-
-fn calc_distance(t1:u64, t2:u64,t3:u64,t4:u64) -> f64 {
-	let f: f64 = 1.0/499200000.0/128.0; // DW3000 frequency (Hz)
-    let s_light: f64 = 299792458.0; // light speed 10⁹m/s
-
-    let tround: f64 = (t4 - t1) as f64;
-    let treply: f64 = (t3 - t2) as f64;
-
-	let tick_tof: f64 = (tround-treply) / 2 as f64;
-
-	let tof: f64 = tick_tof * f ;
-
-	s_light * tof // (meters)
 }
